初期化の改善と不要な引数の削除
`DriverBase.h` で `HwVertexBufferInfo` の `padding` 配列にデフォルト初期化を追加し、常にゼロで初期化されるようにしました。
`HandleAllocator.cpp` と `HandleAllocator.h` では、使用されていない引数 `size` と `age` を削除し、コードをクリーンにしました。また、`handleToPointer` メソッド内で `offset` の型を `size_t` に変更しました。
`OpenGLDriver.cpp` では、`glDrawElements` の `indexOffset` 引数を `uintptr_t` にキャストし、ポインタの安全性を向上させました。さらに、`updateVertexArrayObject` メソッド内で `index` の型を `GLuint` にキャストしました。
`OpenGLDriver.h` では、`GLIndexBuffer` の `usage` メンバーにデフォルト値 `BufferUsage::STATIC` を追加しました。

+++ engine/src/renderer/backend/opengl/OpenGLDriver.cpp
@@ -211,7 +211,7 @@ void OpenGLDriver::draw(PipelineState state, RenderPrimitiveHandle rph, const ui
     GLRenderPrimitive* rp = handle_cast<GLRenderPrimitive*>(rph);
 
     glDrawElements(GLenum(rp->type), static_cast<GLsizei>(indexCount), rp->gl.getIndicesType(),
-                   reinterpret_cast<const void*>(indexOffset));
+                   reinterpret_cast<const void*>(static_cast<uintptr_t>(indexOffset)));
 }
 
 void OpenGLDriver::updateVertexArrayObject(GLRenderPrimitive* rp, GLVertexBuffer* vb)
@@ -227,11 +227,11 @@ void OpenGLDriver::updateVertexArrayObject(GLRenderPrimitive* rp, GLVertexBuffer
         const uint8_t buffer = attribute.buffer;
         if (buffer != Attribute::BUFFER_UNUSED) {
             glBindBuffer(GL_ARRAY_BUFFER, vb->gl.id);
-            const GLuint index = i;
+            const GLuint index = static_cast<GLuint>(i);
             const GLint size = static_cast<GLint>(OpenGLUtility::getComponentCount(attribute.type));
             const GLenum type = OpenGLUtility::getComponentType(attribute.type);
             const GLsizei stride = attribute.stride;
-            const void* pointer = reinterpret_cast<void*>(attribute.offset);
+            const void* pointer = reinterpret_cast<void*>(static_cast<uintptr_t>(attribute.offset));
 
             glVertexAttribPointer(index, size, type, GL_FALSE, stride, pointer);
             glEnableVertexAttribArray(index);
